Updated Torizo and Golden Torizo
Torizo now has a "tell" for whether his ranged claw attack will be high or low. His arm now does not raise quite as high when doing a low attack, and the default anim is used for the high attack.

Also updated his head to properly transfer damage done to it to the body when using melee weapons.

Buffed Golden Torizo. He is now tankier, and his mouth bomb attack is faster.