Hunter Emblem added
Wall of Flesh now drops the Hunter Emblem, though it is in addition to his regular emblem drop. In order to properly add it to his pool of emblem drops, I'd have to do a bunch of code workarounds. For now it's fine, but may be updated in the future.