Localization Overhaul Part 1.6 - UA Bandaid
While all that updating stuff was happening, a new stat for UA usage reduction got added, which seems to be roughly parallel with Overheat. I've slapped a bandaid label on the overheat cost reduction for now which should get everything that currently has a UA usage reduction labelled as such and I'll look into a more permanent adjustment later.