Suit Addon Bandaid
I tried to make texture grabbing automatic for suit addons but it made non-Suits shit themselves. I have made it so all of the offending textures automatically route to the Energy Tank item sprite but later I am going to have to go into SuitAddonLoader and fix it for real.