Addonening - Wave Beam Preliminary Implementation
Since Wave will be next, I added a functional skeleton of it into the game, able to be used and stacked as any addon, as well as possessing its own stat modifiers, but lacking its unique properties.
It took all of 5 minutes to do this.
While it was helped by already having all of the assets in the right places with the right names, it still took me like 5 minutes to make from scratch. I feel like that ought to count for something.

Other notes:
-Updated todo
-Re-added recipes
-The keyword system now has names for each layer: the lower one (haha funny fusion dna beam) is the static combo, while the upper one (haha funny charged shot) is the dynamic combo.

Important known issues:
-Shot textures do not properly load the first time they're called in-game. They work as intended on the second and beyond.