Addonening - Wave Beam Fully Functional*!?
Wave Beam now does the two-shots-on-charge thing.

Renamed SpecialComboGet and the matching array edgeCaseData to ComboVisualsGet and visualData, respectively.

Added a new Loader method, EdgeCaseStacker, with its own array that now holds the name of edgeCaseStuff, whose sole function is pretty much just to let Wave Beam do the two-shot thing. It's not hardcoded but I cannot think of a single other thing it'd be good for.

Beam Shots will no longer color themselves if the shapepriority and colorpriority are the same.
On a related note, swapped out the greyscale shot sprites on Charge and Ice for colored ones.

Updated PED Suit sprites, because it bugged me.

Updated TODO.txt.