Total UI Overhaul Part 1
-Two new UI elements
UIPanelTileBackground: Functionally identical to the standard panel but allows for repeating tiled patterns, which has been used to recreate the grid pattern seen in the Super Metroid suit screen.
UIDraggableBase: An (upgraded) draggable UI panel that is invisible and can only be dragged if doing so from one of its child objects. Used to keep elements that are connected but not appended together when the UI is being dragged around.
-Used these two new elements to cut back on bespoke assets for the Arm Cannon UI.
-Lines and wireframe are no longer separate UIElements, but rather UIImages.
-Dynamic-length slot labels, useful for localization.
-Added buttons (with placeholder textures) for soon-to-be-added array slot tabs.
-Added... I guess pseudo-dictionaries to the beam and missile AddonSlotID classes.

Next stop is gonna be the arrays. Oh boy.
...Maybe I'll procrastinate by updating the other UIs.